+++ src/test/java/edu/illinois/library/cantaloupe/image/DimensionTest.java
@@ -215,7 +215,9 @@ public class DimensionTest extends BaseTest {
 
     @Test
     void testToString() {
-        assertEquals("1000x800", instance.toString());
+        instance.setWidth(1000.5);
+        instance.setHeight(800.5);
+        assertEquals("1000.5x800.5", instance.toString());
     }
 
 }