+++ src/test/java/edu/illinois/library/cantaloupe/resource/iiif/ImageResourceTester.java
@@ -331,6 +331,42 @@ public class ImageResourceTester extends ImageAPIResourceTester {
         }
     }
 
+    public void testContentDispositionHeaderSetToInline(URI uri)
+            throws Exception {
+        Client client = newClient(uri);
+        try {
+            Response response = client.send();
+            assertEquals("inline; filename=" + IMAGE + ".jpg",
+                    response.getHeaders().getFirstValue("Content-Disposition"));
+        } finally {
+            client.stop();
+        }
+    }
+
+    public void testContentDispositionHeaderSetToAttachment(URI uri)
+            throws Exception {
+        Client client = newClient(uri);
+        try {
+            Response response = client.send();
+            assertEquals("attachment; filename=" + IMAGE + ".jpg",
+                    response.getHeaders().getFirstValue("Content-Disposition"));
+        } finally {
+            client.stop();
+        }
+    }
+
+    public void testContentDispositionHeaderSetToAttachmentWithFilename(URI uri,
+                                                                        String filename) throws Exception {
+        Client client = newClient(uri);
+        try {
+            Response response = client.send();
+            assertEquals("attachment; filename=" + filename,
+                    response.getHeaders().getFirstValue("Content-Disposition"));
+        } finally {
+            client.stop();
+        }
+    }
+
     public void testLessThanOrEqualToMaxScale(URI uri) {
         Configuration config = Configuration.getInstance();
         config.setProperty(Key.MAX_SCALE, 1.0);